+++ network_dhcp.sh
@@ -5,5 +5,7 @@
 # MIT License
 # Copyright (c) 2015 Alexander Williams, Unscramble <license@unscramble.jp>
 
-/usr/share/udhcpc/default.script deconfig
-/etc/init.d/dhcp.sh
+pid="/var/run/udhcpc.${interface}.pid"
+
+/opt/udhcpc.script deconfig
+/sbin/udhcpc -b -i $interface -h $(/bin/hostname) -p $pid -s /opt/udhcpc.script >/dev/null 2>&1 &